requirements.txt
reflex==0.4.8.post1
google-generativeai
python-dotenv
PyPDF2
